refactor(SignInForm, SignUpForm): move `schema` and related types to separate files
