+++ packages/client/src/widgets/SignInForm/types.ts
@@ -1,0 +1,4 @@
+import type * as yup from 'yup'
+import type schema from './schema'
+
+export type TUseForm = yup.InferType<typeof schema>
